requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
lightgbm>=4.0
catboost>=1.2
shap>=0.45
fastapi>=0.110
uvicorn>=0.29
streamlit>=1.34
networkx>=3.0
joblib>=1.3
PyYAML>=6.0
